requirements.txt
Flask==3.0.2
requests==2.31.0
tkcalendar==1.6.1
pyinstaller
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
flask-cors==4.0.0
waitress==3.0.0 